fix(build): fix all 3 remaining MSVC/GCC/Clang errors
- MetricsRegistry.hpp: add missing <memory> for std::unique_ptr
- MetricsRegistry: change buckets_ from vector<pair<uint64_t,atomic<uint64_t>>>
  to vector<HistogramBucket> — std::atomic is non-movable and cannot live in a
  reallocating std::vector via emplace_back. Fix: resize() + index assignment.
- Scheduler.hpp: change scheduleGuarded to accept any callable F via template
  parameter (was std::function<void(shared_ptr<T>)>). Bare lambdas now deduce
  correctly on MSVC, GCC, and Clang without explicit wrapping at call sites.
- Session.cpp: remove now-redundant std::function<> wrapping at all 3
  scheduleGuarded call sites (connect/delay/idle timers).
- main.cpp: add static_cast<uint16_t> for all stoi() -> uint16_t assignments
  (MSVC C4244 narrowing warning treated as error via /WX)

+++ src/main.cpp
@@ -49,11 +49,11 @@ namespace {
     bool parseHostPort(const std::string& arg, std::string& host, uint16_t& port) {
         auto pos = arg.find(':');
         if (pos == std::string::npos) {
-            port = std::stoi(arg);
+            port = static_cast<uint16_t>(std::stoi(arg));
             return true;
         }
         host = arg.substr(0, pos);
-        port = std::stoi(arg.substr(pos + 1));
+        port = static_cast<uint16_t>(std::stoi(arg.substr(pos + 1)));
         return true;
     }
 }
@@ -80,7 +80,7 @@ int main(int argc, char* argv[]) {
             parseHostPort(argv[++i], config.upstreamHost, config.upstreamPort);
         }
         else if (arg == "--control" && i + 1 < argc) {
-            config.controlPort = std::stoi(argv[++i]);
+            config.controlPort = static_cast<uint16_t>(std::stoi(argv[++i]));
         }
         else if (arg == "--seed" && i + 1 < argc) {
             config.globalSeed = std::stoull(argv[++i]);
